AnnounceDataのディレクトリを変更
