--- src/components/Annoucements/AnnouncementsData.tsx
@@ -1,167 +1,0 @@
-import { Announcement } from "@/types/types";
-
-export const announcements: Announcement[] = [
-  {
-    title: (
-      <p>
-        <a href="./whats-new" target="_blank" aria-label="What's new?ページ">
-          What&apos;s new?ページ
-        </a>
-        を公開しました!
-      </p>
-    ),
-    content: (
-      <p>
-        <a href="./whats-new" target="_blank" aria-label="What's new?ページ">
-          What&apos;s new?ページ
-        </a>
-        に62回工大祭ホームページの変更点をまとめました!
-      </p>
-    ),
-    date: "2024-11-11",
-  },
-  {
-    title: (
-      <p>
-        諸事情により2日目後夜祭のラストを飾るプロジェクションマッピングは中止となりました
-      </p>
-    ),
-    date: "2024-11-11",
-    category: "important",
-  },
-  {
-    title: <p>工大祭Webサイトをリニューアルしました!</p>,
-    content: (
-      <p>
-        工大祭Webサイトをリニューアルし、新しいロード画面の追加や利便性・パフォーマンスが向上しました!
-      </p>
-    ),
-    date: "2024-11-01",
-  },
-  {
-    title: (
-      <p>
-        <a href="./project/" target="_blank" aria-label="企画ページ">
-          企画情報
-        </a>
-        を公開しました
-      </p>
-    ),
-    date: "2024-11-01",
-  },
-  {
-    title: (
-      <p>
-        <a href="./project/tyuya/" target="_blank" aria-label="中夜祭">
-          中夜祭のページ
-        </a>
-        を更新しました
-      </p>
-    ),
-    date: "2024-10-15",
-  },
-  {
-    title: (
-      <p>
-        <a
-          href="./project/seiyu/"
-          target="_blank"
-          aria-label="声優トークショー"
-        >
-          声優トークショーのページ
-        </a>
-        を更新しました
-      </p>
-    ),
-    date: "2024-10-10",
-  },
-  {
-    title: (
-      <p>
-        <a
-          href="https://x.com/nitfes/status/1840949759592558721"
-          target="_blank"
-          aria-label="パンフレットコンテスト"
-        >
-          パンフレットコンテストの結果
-        </a>
-        を発表しました
-      </p>
-    ),
-    date: "2024-10-01",
-  },
-  {
-    title: (
-      <p>
-        <a href="./pfc/" target="_blank" aria-label="パンフレットコンテスト">
-          パンフレットコンテスト決勝戦
-        </a>
-        を更新しました
-      </p>
-    ),
-    date: "2024-07-28",
-  },
-  {
-    title: (
-      <p>
-        <a href="./pfc/" target="_blank" aria-label="パンフレットコンテスト">
-          パンフレットコンテストのページ
-        </a>
-        を公開しました
-      </p>
-    ),
-    date: "2024-06-30",
-  },
-  {
-    title: (
-      <p>
-        <a
-          href="./project/spark/"
-          target="_blank"
-          aria-label="学生ステージ企画"
-        >
-          学生ステージ企画
-        </a>
-        の募集を開始しました
-      </p>
-    ),
-    date: "2024-06-01",
-  },
-  {
-    title: <p>水無月杯の延期が決定しました</p>,
-    content: (
-      <p>
-        水無月杯の延期が決定しました
-        <br />
-        水無月杯で開催予定だったソフトボール部門の中止が決定しました
-      </p>
-    ),
-    date: "2024-04-08",
-  },
-  {
-    title: (
-      <p>
-        <a href="./minaduki/" target="_blank" aria-label="水無月杯">
-          水無月杯のページ
-        </a>
-        を公開しました
-      </p>
-    ),
-    date: "2024-03-19",
-  },
-  {
-    title: (
-      <p>
-        <a href="./sinkan/" target="_blank" aria-label="新歓">
-          新歓のページ
-        </a>
-        を公開しました
-      </p>
-    ),
-    date: "2024-03-08",
-  },
-  {
-    title: <p>掲載情報を第62回に更新しました</p>,
-    date: "2024-03-08",
-  },
-];
